split `Reports` component into smaller components
https://issues.redhat.com/browse/AA-196
