+++ src/Containers/Reports/Details/ReportTable/index.tsx
@@ -1,0 +1,3 @@
+import ReportTable from './ReportTable';
+
+export default ReportTable;
