+++ src/Containers/Reports/Shared/types.ts
@@ -12,7 +12,7 @@ export interface ReportGeneratorParams {
   defaultParams: Params;
   expandRows?: boolean;
   extraAttributes: AttributesType;
-  listAttributes?: Array<any>,
+  listAttributes?: string[],
   readData: (options: ParamsWithPagination) => Promise<ApiJson>;
   readOptions: (options: Params) => Promise<ApiJson>;
   schemaFnc: SchemaFnc;
